comments
changed tests
changed hash of block header
